requirements.txt
Django==5.2.8
sqlparse==0.5.3
django-environ==0.12.0
psycopg2-binary==2.9.11
djangorestframework==3.16.1
redis==6.4.0
drf-yasg==1.21.10
gunicorn==23.0.0
whitenoise==6.11.0
django-filter==25.2
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-storages==1.14.6
Pillow==12.1.0
openpyxl==3.1.5
